Revert "Restructured project to MVC"
