+++ lib/admin/logic/add_details.dart
