+++ lib/admin/logic/get_details.dart
