+++ lib/admin/widgets/admin_bottomnavbar.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:student_app/view/admin/admin_screens.dart';
-import 'package:student_app/view/admin/screens/mark_entry.dart';
-import 'package:student_app/view/common/settings.dart';
+import 'package:student_app/admin/screens/mark_entry.dart';
+import 'package:student_app/admin/screens/upload_notes.dart';
+import 'package:student_app/admin/screens/upload_profile.dart';
+import 'package:student_app/common/settings.dart';
 
 import 'dialog.dart';
