+++ lib/admin/widgets/designs.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:student_app/controller.dart';
 
+import '../logic/add_details.dart';
+import '../logic/attendance.dart';
 import 'dialog.dart';
 
 class Designs extends StatefulWidget {
