+++ lib/common/onboarding_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:introduction_screen/introduction_screen.dart';
-import 'package:student_app/view/common/common_screens.dart';
+import 'package:student_app/common/login.dart';
 
 /// onBoarding page UI
 class OnBoardingPage extends StatefulWidget {
