--- lib/controller.dart
@@ -1,4 +1,0 @@
-export 'controller/add_details.dart';
-export 'controller/attendance.dart';
-export 'controller/get_details.dart';
-export 'controller/process_data.dart';
