--- lib/models.dart
@@ -1,1 +1,0 @@
-export 'models/db_model.dart';
