--- lib/view/admin/admin_screens.dart
@@ -1,3 +1,0 @@
-export 'screens/mark_entry.dart';
-export 'screens/upload_notes.dart';
-export 'screens/upload_profile.dart';
