--- lib/view/admin/admin_widgets.dart
@@ -1,3 +1,0 @@
-export 'widgets/admin_bottomnavbar.dart';
-export 'widgets/designs.dart';
-export 'widgets/dialog.dart';
