--- lib/view/common/common_screens.dart
@@ -1,3 +1,0 @@
-export 'login.dart';
-export 'onboarding_screen.dart';
-export 'settings.dart';
