--- lib/view/student/student_screens.dart
@@ -1,4 +1,0 @@
-export 'screens/dashboard.dart';
-export 'screens/grade.dart';
-export 'screens/profile.dart';
-export 'screens/view_notes.dart';
