--- lib/view/student/student_widgets.dart
@@ -1,1 +1,0 @@
-export '../student/widgets/student_bottomnavbar.dart';
